+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+
+## 3.1.0 - 2022-08-16
+
+### Changed
+- Penalties are sent as part of the Session API call, to avoid duplicate Sessions and duplicate Penalties
+- Sentry only runs on currently maintained game versions
+
+### Added
+- Sessions now also send the UDP season_identifier to allow for more reliable season assignment
+
+
 ## 3.0.4 - 2022-07-19
 
 ### Fixed